Support migration for CI
