+++ .travis.yml
@@ -23,4 +23,4 @@ deploy:
     branch: qa
 after_deploy:
 - ssh -i ~/.ssh/key/efcsydney.pem ec2-user@54.206.106.116 "cd /opt/efcsydney-roster;
-  . ~/.bash_profile; nvm use 8; yarn"
+  . ~/.bash_profile; nvm use 8; yarn; yarn db-migrate"
