--- db/migrate/20171227012505-alter-position-table.js
@@ -1,17 +1,0 @@
-'use strict';
-
-module.exports = {
-  up: (queryInterface, Sequelize) => {
-    return queryInterface.addColumn(
-      'positions',
-      'serviceId',
-      Sequelize.STRING,
-      {
-        after: 'order'
-      }
-    );
-  },
-  down: (queryInterface, Sequelize) => {
-    return queryInterface.removeColumn('positions', 'serviceId');
-  }
-};
