--- db/seed/20171227015336-initialize-services-table.js
@@ -1,16 +1,0 @@
-module.exports = {
-  up: (queryInterface, Sequelize) => {
-    return queryInterface.bulkInsert(
-      'services',
-      [
-        { id: 1, name: 'english', locale: 'en-AU' },
-        { id: 2, name: 'chinese', locale: 'zh-TW' }
-      ],
-      {}
-    );
-  },
-
-  down: (queryInterface, Sequelize) => {
-    return queryInterface.bulkDelete('services', null, {});
-  }
-};
